src/requirements.txt
Flask==2.1.0
Werkzeug==2.2.2
gunicorn==20.1.0
streamlit
requests 
openai
pandas
PyPDF2
google-cloud-storage
google-cloud-bigtable